services/dataset_worker/pyproject.toml
[project]
name = "dataset-worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pika",
    "python-dotenv>=1.1.1",
    "returns>=0.25.0",
]

[tool.uv]
dev-dependencies = ["pyright", "pytest", "ruff"]
